vdev Cashtrail.Banking.Currencies

The Currencies context manages the currencies data of one entity.

See Cashtrail.Banking.Currencies.Currency to have more info about what currencies mean in the application.

Types

currency()

currency() :: Cashtrail.Banking.Currencies.Currency.t()

Functions

change_currency(currency)

change_currency(currency()) :: Ecto.Changeset.t(currency())

Returns an %Ecto.Changeset{} for tracking currency changes.

Expected Arguments

currency - The %Cashtrail.Banking.Currencies.Currency{} to be tracked.

Examples

iex> change_currency(currency)
%Ecto.Changeset{source: %Currency{}}

create_currency(entity, attrs)

create_currency(Cashtrail.Entities.Entity.t(), map()) ::
  {:ok, currency()} | {:error, Ecto.Changeset.t(currency())}

Creates a currency.

Expected Arguments

entity - The %Cashtrail.Entities.Entity{} that the currency references.

params - A map with the params of the currency to be created:

:description (required) - A string that is the description of the currency.

:type - A string that is the type of currency. It can receive "money", "digital_currency", "miles", "cryptocurrency" or "other". Defaults to "money".

:iso_code - A string that is the ISO 4217 code of the currency. Must be unique for the entity and have the exact 3 characters.

:symbol - A string that is the symbol of the currency.

:format - A string that represents the format of the currency. The "%s" refers to the :symbol field, and the "%n" refers to the number. Defaults to "%s%n".

:precision - A integer that represents how much decimal places the currency has. Defaults to 0.

:separator - A string that is used to separate the integer part from the fractional part of the currency. It must have an exact one character or be empty. Defaults to ".".

:delimiter - A string that is used to separate the thousands parts of the currency. Defaults to ".".

:active - A boolean that says if the currency is active and should be displayed in lists of the application. Defaults to true.

See Cashtrail.Banking.Currencies.Currency to have more detailed info about the fields.

Returns

{:ok, %Cashtrail.Banking.Currencies.Currency{}} in case of success.

{:error, %Ecto.Changeset{}} in case of error.

Examples

iex> create_currency(entity, %{field: value})
{:ok, %Currency{}}

iex> create_currency(entity, %{field: bad_value})
{:error, %Ecto.Changeset{}}

delete_currency(currency)

delete_currency(currency()) ::
  {:ok, currency()} | {:error, Ecto.Changeset.t(currency())}

Deletes a currency.

Expected Arguments

currency - The %Cashtrail.Banking.Currencies.Currency{} to be deleted.

Returns

{:ok, %Cashtrail.Banking.Currencies.Currency{}} in case of success.

{:error, %Ecto.Changeset{}} in case of error.

Examples

iex> delete_currency(currency)
{:ok, %Currency{}}

iex> delete_currency(currency)
{:error, %Ecto.Changeset{}}

get_currency!(entity, id)

get_currency!(Cashtrail.Entities.Entity.t(), Ecto.UUID.t() | String.t()) ::
  currency()

Gets a single currency.

Raises Ecto.NoResultsError if the Currency does not exist.

See Cashtrail.Banking.Currencies.Currency to have more detailed info about the returned struct.

Expected Arguments

entity - The %Cashtrail.Entities.Entity{} that the currency references.

id - A string that is the unique id of the currency to be found.

Examples

iex> get_currency!(entity, 123)
%Currency{}

iex> get_currency!(entity, 456)
** (Ecto.NoResultsError)

list_currencies(entity, options \\ [])

list_currencies(Cashtrail.Entities.Entity.t(), keyword()) ::
  Cashtrail.Paginator.Page.t()

Returns a %Cashtrail.Paginator.Page{} struct with a list of currencies in the :entries field.

If no currencies are found, return an empty list in the :entries field.

Expected arguments

entity - The %Cashtrail.Entities.Entity{} that the currency references.

options - A keyword list of the following options:

:filter - filters by following attributes:

:type or "type"

:active or "active"

:search - search currencies by :description, :iso_code and :symbol.

See Cashtrail.Paginator.paginate/2 to know about the pagination options.

See Cashtrail.Banking.Currencies.Currency to have more detailed info about each field to be filtered or searched.

Examples

iex> list_currencies(entity)
%Cashtrail.Paginator.Page{entries: [%Currency{}, ...], ...}

iex> list_currencies(entity, page: 2)
%Cashtrail.Paginator.Page{entries: [%Currency{}, ...], page: 2}

iex> list_currencies(entity, filter: %{type: "money"})
%Cashtrail.Paginator.Page{entries: [%Currency{type: "money"}, ...]}

iex> list_currencies(entity, filter: %{search: "my"})
%Cashtrail.Paginator.Page{entries: [%Currency{description: "my money"}, ...]}

update_currency(currency, attrs)

update_currency(currency(), map()) ::
  {:ok, currency()} | {:error, Ecto.Changeset.t(currency())}

Updates a currency.

Expected Arguments

currency - The %Cashtrail.Banking.Currencies.Currency{} to be updated.

params - A map with the field of the currency to be updated. See create_currency/2 to know about the params that can be given.

Returns

{:ok, %Cashtrail.Banking.Currencies.Currency{}} in case of success.

{:error, %Ecto.Changeset{}} in case of error.

Examples

iex> update_currency(currency, %{field: new_value})
{:ok, %Currency{}}

iex> update_currency(currency, %{field: bad_value})
{:error, %Ecto.Changeset{}}