vdev Cashtrail.Banking.Currencies.Currency

This is an Ecto.Schema struct that represents a currency of the entity.

Warning: Don't use the functions of this module. Only use this module as a struct to represent a currency. The functions of this module are internal and can change over time. Only manipulate currencies through the Cashtrail.Banking.Currencies that is the context for this.

Definition

According to Investopedia, the currency is any form when in use or circulation as a medium of exchange. This can be any system of money in common use by people.

The common examples of currencies are Brazilia, reais (R$), U.S. dollars (US$), euros (€), Japanese yen (¥), and pounds sterling (£). There are cryptocurrencies too. In this application, you don't have to be stuck with that definition. You can register any means of exchange you use as a currency. It could be your airline miles, some other kind of currency that you use in any community.

Anything you want to track can be registered as a currency. Just know that a bank account can only have one currency registered. And all transactions will be performed in that currency. The only way to convert a currency to another is through a special form of transaction called 'exchange'. So you can use this to track your expenses in another currency on one travel or track your miles usage.

Each entity will have their currencies, then you can register or delete currencies without worries. So you are free to define how you will use this feature.

Fields

:id - The unique id of the currency.

:description - The description of the currency.

:type - The type of currency. Can be:

"money" - ordinary currencies like dollars, euro, yen, etc. This is the default value if no type is chosen.

"cryptocurrency" - digital currencies that use cryptographical functions to conduct financial transactions.

"virtual" - unregulated digital currencies, used and accepted among the members of a specific virtual community. For example loyalty points, game points, etc.

"other" - other types of currencies that don't match the previous categories.

:iso_code - The ISO 4217 code of the currency.

:active - Says if the currency is active or not.

:symbol - The symbol of the currency, like R$, US$, €, ¥, or £ for example.

:precision - Every currency can have a different number of decimal places. For example, the dinar has three decimal places, dollar two, and yen zero. This field can be used to help round and format the amounts for the currency correctly.

:separator - The field can be used to separate the integer part from the fractional part of the currency.

:delimiter - The field can be used to separate the thousands parts of the currency.

:format - This field can be used to know in what format the symbol and the number should be displayed. The "%s" represents the symbol, and the "%n" represents the number. So, if you format 100 dollars using "%s %n" the expected format will be "US$ 100.00". This field, as the :precision, :separator and :delimiter, only brings a reference to be used by the libraries that will perform the currency formating.

:inserted_at - When the currency was inserted at the first time.

:updated_at - When the currency was updated at the last time.

See Cashtrail.Banking.Currencies to know how to list, get, insert, update, and delete currencies.

Types

t()

t() :: %Cashtrail.Banking.Currencies.Currency{
  __meta__: Ecto.Schema.Metadata.t(),
  active: boolean() | nil,
  delimiter: String.t() | nil,
  description: String.t() | nil,
  format: String.t() | nil,
  id: Ecto.UUID.t() | nil,
  inserted_at: NaiveDateTime.t() | nil,
  iso_code: String.t() | nil,
  precision: integer() | nil,
  separator: String.t() | nil,
  symbol: String.t() | nil,
  type: String.t() | nil,
  updated_at: NaiveDateTime.t() | nil
}