vdev Cashtrail.Contacts.Address

This is an Ecto.Schema struct that represents an address of the contact.

Warning: Don't use the functions of this module. Only use this module as a struct to represent an address. The functions of this module are internal and can change over time. Only manipulate addresses through the Cashtrail.Contacts that is the context for this.

The address is stored in a json field of the contact, and not in a specific table in the database. This struct is used to force the address to a specific schema in this JSON field.

Each country has its type of address. So, no fields are required. You can choose to allow only to fill :line_1 and :line_2 in your frontend, or allow fill the other fields instead.

Fields

:id - The unique id of the address.

:street - street part of the address.

:number - number part of the address.

:complement - complement part of the address, like apartment number for instance.

:district - district part of the address.

:city - city part of the address.

:state - state or province part of the address. This depends on the country.

:country - The country of the address.

:zip - The zip code of the address. This field is not validated, so you can insert whatever the zip code of any country you want.

:line_1 - Line 1 can have the street and number in some countries (like in the US).

:line_2 - Line 2 can have the city, state, and zip code in some countries (like in the US).

See Cashtrail.Contacts.create_contact/2 to know how to create a contact with an address, or Cashtrail.Contacts.create_contact/2 to know how to update an address of a contact, or insert a new one.

Types

t()

t() :: %Cashtrail.Contacts.Address{
  city: String.t() | nil,
  complement: String.t() | nil,
  country: String.t() | nil,
  district: String.t() | nil,
  id: Ecto.UUID.t() | nil,
  line_1: String.t() | nil,
  line_2: String.t() | nil,
  number: String.t() | nil,
  state: String.t() | nil,
  street: String.t() | nil,
  zip: String.t() | nil
}