vdev Cashtrail.Entities

The Entities context manages the data related to entities. An Entity keeps all financial data of something, that can be a company, financial finances, organization, church, event, etc. And they can have one owner or other members, as well.

See Cashtrail.Entities.Entity to have more info about entity.

Types

entity()

entity() :: Cashtrail.Entities.Entity.t()

entity_member()

entity_member() :: Cashtrail.Entities.EntityMember.t()

Functions

add_member(entity, user, permission \\ "read")

add_member(entity(), Cashtrail.Users.User.t(), String.t()) ::
  {:ok, entity_member()} | {:error, :invalid | Ecto.Changeset.t(entity())}

Add a user as an entity_member for the entity giving permission.

Expected Arguments

entity - The %Cashtrail.Entities.Entity{} that the member will be added.

user - A %Cashtrail.Users.User{} that is the user to be added as a member. The user cannot be the owner of the entity, otherwise, it will return an error.

permission - A string with the permission that will be given to the member. It can be: "read", "write" or "admin".

See Cashtrail.Entities.EntityMember to have more detailed info about the permissions.

Returns

{:ok, %Cashtrail.Entities.EntityMember{}} in case of success.

{:error, :invalid} in case of the user be the owner of the entity.

{:error, %Ecto.Changeset{}} in case of error.

Examples

iex> add_member(entity, user)
{:ok, %EntityMember{}}

iex> add_member(entity, invalid_user)
{:error, %Ecto.Changeset{}}

belongs_to?(entity, user)

belongs_to?(entity(), Cashtrail.Users.User.t()) :: boolean()

Returns a boolean that says if the entity belongs to the user.

Expected Arguments

user - The %Cashtrail.Users.User{} to be deleted.

Examples

iex> belongs_to?(%Entity{owner_id: "aaa"}, %User{id: "aaa"})
true

iex> belongs_to?(%Entity{owner_id: "bbb"}, %User{id: "aaa"})
false

change_entity(entity)

change_entity(entity()) :: Ecto.Changeset.t(entity())

Returns an %Ecto.Changeset{} for tracking entity changes.

Expected Arguments

entity - The %Cashtrail.Entities.Entity{} to be tracked.

Examples

iex> change_entity(entity)
%Ecto.Changeset{source: %Entity{}}

change_member(entity_member)

change_member(entity_member()) :: Ecto.Changeset.t(entity_member())

Returns an %Ecto.Changeset{} for tracking entity_member changes.

Expected Arguments

entity_member - The %Cashtrail.Entities.EntityMember{} to be tracked.

Examples

iex> change_member(entity_member)
%Ecto.Changeset{source: %EntityMember{}}

create_entity(user, attrs, create_tenants \\ true)

create_entity(Cashtrail.Users.User.t(), map(), boolean()) ::
  {:ok, entity()} | {:error, Ecto.Changeset.t(entity())}

Creates an entity.

Expected Arguments

params - A map with the params of the entity to be created:

:name (required) - A string with the name or description of the entity.

:type - A string with the type of the entity. It can be "personal", "company" or "other". Defaults to "personal".

:owner_id - A string that references to the Cashtrail.Users.User that is the owner of the entity.

See Cashtrail.Entities.Entity to have more detailed info about the fields.

Returns

{:ok, %Cashtrail.Entities.Entity{}} in case of success.

{:error, %Ecto.Changeset{}} in case of error.

Examples

iex> create_entity(user, %{field: value})
{:ok, %Entity{}}

iex> create_entity(user, %{field: bad_value})
{:error, %Ecto.Changeset{}}

create_member(entity, attrs)

create_member(entity(), map()) ::
  {:ok, entity_member()} | {:error, Ecto.Changeset.t(entity_member())}

Creates an entity_member for the entity.

Expected Arguments

entity - The %Cashtrail.Entities.Entity{} that the member will be created.

params - A map with the params of the user to be created:

:permission (required) - a string with the permission that will be given to the member. It can be: "read", "write" or "admin".

:user_id - A string with a reference to one Cashtrail.Users.User to be added as a member to the entity.

:user - A map of the Cashtrail.Users.User that should be created as a member of the entity. See Cashtrail.Users.create_user/1 docs to know more about the accepted params.

See Cashtrail.Entities.EntityMember to have more detailed info about the fields.

Returns

{:ok, %Cashtrail.Entities.EntityMember{}} in case of success.

{:error, %Ecto.Changeset{}} in case of error.

Examples

iex> create_member(entity, %{field: value})
{:ok, %EntityMember{}}

iex> create_member(entity, %{field: bad_value})
{:error, %Ecto.Changeset{}}

delete_entity(entity, drop_tenants \\ true)

Deletes an entity.

Expected Arguments

entity - The %Cashtrail.Entities.Entity{} to be deleted.

Returns

{:ok, %Cashtrail.Entities.Entity{}} in case of success.

{:error, %Ecto.Changeset{}} in case of error.

Examples

iex> delete_entity(entity)
{:ok, %Entity{}}

iex> delete_entity(entity)
{:error, %Ecto.Changeset{}}

get_entity!(id)

get_entity!(Ecto.UUID.t()) :: entity()

Gets a single entity.

Raises Ecto.NoResultsError if the Entity does not exist.

See Cashtrail.Entities.Entity to have more detailed info about the returned struct.

Expected Arguments

id - A string that is the unique id of the entity to be found.

Examples

iex> get_entity!(123)
%Entity{}

iex> get_entity!(456)
** (Ecto.NoResultsError)

get_member_permission(entity, user)

get_member_permission(entity(), Cashtrail.Users.User.t()) :: atom()

Returns the member permission as an atom or :unauthorized if the member is not found. If the user is the owner, return permission as :admin.

Expected Arguments

entity - The %Cashtrail.Entities.Entity{} that the member belongs.

user - The %Cashtrail.Users.User{} to know the permission.

Examples

iex> get_member_permission(entity, user)
:admin

iex> get_member_permission(entity, another_user)
:unauthorized

list_entities(options \\ [])

list_entities(keyword()) :: Cashtrail.Paginator.Page.t(entity())

Returns a %Cashtrail.Paginator.Page{} struct with a list of entities in the :entries field.

Expected arguments

options - A keyword list of the following options:

:filter => filters by following attributes:

:type or "type"

:status or "status"

:search => search entities by :name.

See Cashtrail.Paginator.paginate/2 to see paginations options.

See Cashtrail.Entities.Entity to have more detailed info about the fields to be filtered or searched.

Examples

iex> list_entities()
%Cashtrail.Paginator.Page{entries: [%Entity{}, ...]}

iex> list_entities(filter: %{type: "company"})
%Cashtrail.Paginator.Page{entries: [%Entity{type: "company"}, ...]}

iex> list_entities(search: "my")
%Cashtrail.Paginator.Page{entries: [%Entity{name: "My company"}, ...]}

list_entities_from(user, params \\ [])

list_entities_from(Cashtrail.Users.User.t(), keyword()) ::
  Cashtrail.Paginator.Page.t(entity())

Returns a %Cashtrail.Paginator.Page{} struct with a list of entities in the :entries field from the given user.

Expected arguments

user - A %Cashtrail.Users.User{} that owns or is member of the entity.

options - A keyword list of the following options:

:filter => filters by following attributes:

:type or "type"

:status or "status"

:search => search entities by :name.

See Cashtrail.Paginator.paginate/2 to see paginations options.

See Cashtrail.Entities.Entity to have more detailed info about the fields to be filtered or searched.

Examples

iex> list_entities_from(owner)
%Cashtrail.Paginator.Page{entries: [%Entity{}, ...]}

iex> list_entities_from(member)
%Cashtrail.Paginator.Page{entries: [%Entity{}, ...]}

list_members(entity, options \\ [])

list_members(entity(), keyword() | map()) ::
  Cashtrail.Paginator.Page.t(entity_member())

Returns a %Cashtrail.Paginator.Page{} struct with a list of entity_members in the :entries field.

Expected arguments

options - A keyword list of the following options:

:filter => filters by following attributes:

:permission or "permission"

:search => search users by its user :name.

See Cashtrail.Paginator.paginate/2 to see paginations options.

See Cashtrail.Entities.EntityMember to have more detailed info about the fields to be filtered or searched.

Examples

iex> list_entity_members(entity)
%Cashtrail.Paginator.Page{entries: [%EntityMember{}, ...]}

iex> list_entity_members(entity, filter: %{permission: "read"})
%Cashtrail.Paginator.Page{entries: [%EntityMember{permission: "read"}, ...]}

iex> list_entity_members(entity, search: "my")
%Cashtrail.Paginator.Page{entries: [%EntityMember{user: %User{name: "My Name"}}, ...]}

member_from_user(entity, user)

member_from_user(entity(), Cashtrail.Users.User.t()) :: entity_member() | nil

Returns the %Cashtrail.Entities.EntityMember{} from the user and the entity. Returns nil if the user is not a member of the entity or if it is the owner.

Expected Arguments

entity - The %Cashtrail.Entities.Entity{} that the member belongs.

user - The %Cashtrail.Users.User{} to have the entity_member found.

Examples

iex> member_from_user(entity, user)
%EntityMember{}

iex> member_from_user(entity, owner)
nil

iex> member_from_user(entity, non_member_user)
nil

remove_member(entity, user)

remove_member(entity(), Cashtrail.Users.User.t()) ::
  {:ok, entity_member()} | {:error, :not_found}

Removes an entity_member from the entity.

Expected Arguments

entity - The %Cashtrail.Entities.Entity{} that the member will be removed.

user - A %Cashtrail.Users.User{} that is the user to be removed as a member of the given entity.

Returns

{:ok, %Cashtrail.Entities.EntityMember{}} in case of success.

{:error, :not_found} if the user is not a member of the entity.

Examples

iex> delete_entity_member(entity_member)
{:ok, %EntityMember{}}

iex> delete_entity_member(entity_member)
{:error, :not_found}

transfer_ownership(entity, from_user, to_user)

transfer_ownership(entity(), Cashtrail.Users.User.t(), Cashtrail.Users.User.t()) ::
  {:error, :unauthorized} | {:ok, entity()}

Transfer the ownership of an entity from one user to another.

Expected Arguments

entity - The %Cashtrail.Entities.Entity{} to be transfered.

from_user - The %Cashtrail.Users.User{} to be transfered.

to_user - The %Cashtrail.Users.User{} to be transfered.

Returns

{:ok, %Entity{}} if the entity is transfered successfully.

{:error, changeset} if to_user is invalid or it's not found.

{:error, :unauthorized} if from_user is not the owner of the entity.

Effects

After the ownership transference, the previous owner (from_user) becomes a member of the entity with :admin permissions.

Examples

iex> transfer_ownership(entity, from_user, to_user)
{:ok, %Entity{}}

iex> transfer_ownership(entity, from_user, to_user)
{:error, %Ecto.Changeset{source: %Entity{}}}

iex> transfer_ownership(entity, invalid_from, to_user)
{:error, :unauthorized}

update_entity(entity, attrs)

update_entity(entity(), map()) ::
  {:ok, entity()} | {:error, Ecto.Changeset.t(entity())}

Updates an entity.

Expected Arguments

user - The %Cashtrail.Entities.Entity{} to be updated.

params - A map with the field of the entity to be updated. See create_entity/2 to know about the params that can be given.

Returns

{:ok, %Cashtrail.Entities.Entity{}} in case of success.

{:error, %Ecto.Changeset{}} in case of error.

Examples

iex> update_entity(entity, %{field: new_value})
{:ok, %Entity{}}

iex> update_entity(entity, %{field: bad_value})
{:error, %Ecto.Changeset{}}

update_member_permission(entity, user, permission)

update_member_permission(entity(), Cashtrail.Users.User.t(), String.t()) ::
  {:ok, entity_member()}
  | {:error, Ecto.Changeset.t(entity_member()) | :invalid | :not_found}

Updates the member's permission.

If the user is not a member or is the owner, it returns an error. The owner will always have admin permission.

Expected Arguments

entity - The %Cashtrail.Entities.Entity{} that the member will have the permissions updated.

user - A %Cashtrail.Users.User{} that is the user to have the permissions updated.

permission - A string with the permission that will be given to the member. It can be: "read", "write" or "admin".

Returns

{:ok, %Cashtrail.Entities.EntityMember{}} in case of success.

{:error, :invalid} if the user is the owner of the entity.

{:error, :not_found} if the user is not a member of the entity.

{:error, %Ecto.Changeset{}} in case of validation errors.

Examples

iex> update_member_permission(entity, user, "write")
{:ok, %EntityMember{}}

iex> update_member_permission(entity, user, "invalid")
{:error, %Ecto.Changeset{}}

iex> update_member_permission(entity, owner, "write")
{:error, :invalid}

iex> update_member_permission(entity, another_user, "write)
{:error, :not_found}