vdev Cashtrail.Entities.EntityMember

This is an Ecto.Schema struct that represents a member of a Cashtrail.Entity that links authorized users to the Cashtrail.Entity, except the owner.

Warning: Don't use the functions of this module. Only use this module as a struct to represent a contact. The functions of this module are internal and can change over time. Only manipulate contacts through the Cashtrail.Entities that is the context for this.

The EntityMember is a member of the entity. As a member of the entity, the user can have permission to read, create and update records, or even admin the entity.

The owner cannot be a member of the Cashtrail.Entity.

Fields

:id - The unique id of the entity member.

:permission - The permission of the entity member. The permissions can be:

"read" - With this permission, the member can read the data from the entity.

"write" - With this permission, the member can read, create, modify, and delete data from the entity, except change the entity settings or manage the members of the entity.

"admin" - With this permission, the member can have all permissions from write, change the settings, and manage the members of the entity.

:entity - The entity that the member is part of, related to Cashtrail.Entities.Entity.

:entity_id - The id of the entity that the member is part of.

:user - The user that is a member of the entity, related to Cashtrail.Users.User.

:user_id - The id of the user that is member of the entity.

:inserted_at - When the entity member was inserted at the first time.

:updated_at - When the entity member was updated at the last time.

See Cashtrail.Entities to know how to list, get, insert, update, and delete entity members.

Types

t()

t() :: %Cashtrail.Entities.EntityMember{
  __meta__: Ecto.Schema.Metadata.t(),
  entity: Ecto.Association.NotLoaded.t() | Cashtrail.Entities.Entity.t() | nil,
  entity_id: Ecto.UUID.t() | nil,
  id: Ecto.UUID.t() | nil,
  inserted_at: NaiveDateTime.t() | nil,
  permission: String.t() | nil,
  updated_at: NaiveDateTime.t() | nil,
  user: Ecto.Association.NotLoaded.t() | Cashtrail.Users.User.t() | nil,
  user_id: Ecto.UUID.t() | nil
}