vdev Cashtrail.Entities.Tenants

Deals with tenants creation for Entity.

Every created Entity should be a tenant and have its data. Tenants are schemas in the Postgres having the data related to the Entity.

Functions

create(entity)

create(%Cashtrail.Entities.Entity{
  __meta__: term(),
  id: Ecto.UUID.t(),
  inserted_at: term(),
  members: term(),
  name: term(),
  owner: term(),
  owner_id: term(),
  status: term(),
  type: term(),
  updated_at: term()
}) :: {:error, String.t()} | {:ok, Ecto.UUID.t()}

Create a tenant for the given Entity.

Expected arguments

A %Cashtrail.Entities.Entity{} struct of the tenant that will be created.

Returns

{:ok, entity} - If the tenant creation was successful performed.

{:error, reason} - In case of errors.

See Triplex.create/2 docs for more information.

drop(entity)

drop(%Cashtrail.Entities.Entity{
  __meta__: term(),
  id: Ecto.UUID.t(),
  inserted_at: term(),
  members: term(),
  name: term(),
  owner: term(),
  owner_id: term(),
  status: term(),
  type: term(),
  updated_at: term()
}) :: {:error, String.t()} | {:ok, Ecto.UUID.t()}

Drop a tenant for the given Entity.

Expected arguments

A %Cashtrail.Entities.Entity{} struct of the tenant that will be dropped.

Returns

{:ok, entity} - If the tenant creation was successful performed.

{:error, reason} - In case of errors.

See Triplex.create/2 docs for more information.

put_prefix(queryable, entity)

put_prefix(Ecto.Queryable.t(), Cashtrail.Entities.Entity.t()) :: Ecto.Query.t()

Return the given Ecto.Queryable with the prefix configured.

Expected arguments

queryable - The Ecto.Queryable that the the prefix will be configured.

A %Cashtrail.Entities.Entity{} struct of the tenant that want to configure the prefix.

See Triplex.to_prefix/1 docs for more information.

to_prefix(entity)

to_prefix(Cashtrail.Entities.Entity.t()) :: String.t()

Return the prefix from Entity.

Expected arguments

A %Cashtrail.Entities.Entity{} struct of the tenant that want to get the prefix.

See Triplex.to_prefix/1 docs for more information.