vdev Cashtrail.QueryBuilder

This module is responsible to generate queries for contexts to filter and search by schema fields.

Functions

build_filter(query, params, allowed_filters)

build_filter(Ecto.Queryable.t(), nil | map(), [atom()]) ::
  Ecto.Query.t() | Ecto.Queryable.t()

Returns a Ecto.Query with the queries based on the given filters and allowed fields, or the given Ecto.Queryable without changes.

The query will use only the params that have the key in the allowed_filters param.

Expected arguments

query - The Ecto.Queryable that the query will be performed.

params - A map keys of the fields and values to be filtered. The keys can be even string or atom, and the values must be the same type of the data in database, or can receive a list with data in the same type of the data in the database.

allowed_fields - A list of atom with the fields that will be used to perform the query. The query will be based only in params that have the keys matching this param.

Examples

iex> Cashtrail.QueryBuilder.build_filter(Cashtrail.Users.User, nil, []) Cashtrail.Users.User

iex> Cashtrail.QueryBuilder.build_filter(Cashtrail.Users.User, %{first_name: "my name"}, [:first_name]) #Ecto.Query

iex> Cashtrail.QueryBuilder.build_filter(Cashtrail.Users.User, %{first_name: ["my", "name"]}, [:first_name]) #Ecto.Query

iex> Cashtrail.QueryBuilder.build_filter(Cashtrail.Users.User, %{"first_name" => "my name"}, [:first_name]) #Ecto.Query

iex> Cashtrail.QueryBuilder.build_filter(Cashtrail.Users.User, %{"first_name" => ["my", "name"]}, [:first_name]) #Ecto.Query

build_search(query, term, fields)

build_search(Ecto.Queryable.t(), nil | String.t(), [atom()]) ::
  Ecto.Query.t() | Ecto.Queryable.t()

Returns a Ecto.Query with the queries based on the given term and fields, or the given Ecto.Queryable without changes.

The search is implement using ILIKE in the fields of the given queryable schema. And the term must be a string.

Expected arguments

query - The Ecto.Queryable that the query will be performed.

term - A string with the text that will be searched.

fields - A list of atom with the fields that will be used to perform the query. The given fields must be string or text, otherwise you will get an error from Ecto.

Examples

iex> Cashtrail.QueryBuilder.build_search(Cashtrail.Users.User, nil, []) Cashtrail.Users.User

iex> Cashtrail.QueryBuilder.build_search(Cashtrail.Users.User, "my name", [:first_name, :last_name]) #Ecto.Query