vdev Cashtrail.Users

The Users context manages the users data of one entity and performs user authentication.

See Cashtrail.Users.User to have more info about user.

Types

user()

user() :: Cashtrail.Users.User.t()

Functions

authenticate(email, password)

authenticate(String.t(), String.t()) ::
  {:ok, user()} | {:error, :not_found | :unauthorized}

Authenticates a user with its email and password.

Expected Arguments

email - A string that is the email of the user.

password - A string that is the expected password of the user.

Returns

{:ok, user} if user is found and the passwords match.

{:error, :unauthorized} if passwords does not match.

{:error, :not_found} if user email is not found.

Examples

iex> authenticate(email, password)
{:ok, %User{}}

iex> authenticate(email, wrong_pass)
{:error, :unauthorized}

iex> authenticate(wrong_email, password)
{:error, :not_found}

change_user(user)

change_user(user()) :: Ecto.Changeset.t(user())

Returns an %Ecto.Changeset{} for tracking user changes.

Expected Arguments

user - The %Cashtrail.Users.User{} to be tracked.

Examples

iex> change_user(user)
%Ecto.Changeset{source: %User{}}

create_user(attrs)

create_user(map()) :: {:ok, user()} | {:error, Ecto.Changeset.t(user())}

Creates a user.

Expected Arguments

params - A map with the params of the user to be created:

:email (required) - A string with the email of the user. Must be a valid email and unique in the application.

:first_name (required) - A string with the first name of the user.

:last_name - A string with the last name of the user.

:password (required) - A string with the password of the user to be created. The password must contain at least one letter, one number, and one special character.

:password_confirmation (required) - A string with password confirmation of the user to be created. Must be the equals the :password field.

See Cashtrail.Users.User to have more detailed info about the fields.

Returns

{:ok, %Cashtrail.Users.User{}} in case of success.

{:error, %Ecto.Changeset{}} in case of error.

Examples

iex> create_user(%{field: value})
{:ok, %User{}}

iex> create_user(%{field: bad_value})
{:error, %Ecto.Changeset{}}

delete_user(user)

delete_user(user()) :: {:ok, user()} | {:error, Ecto.Changeset.t(user())}

Deletes a user.

Expected Arguments

user - The %Cashtrail.Users.User{} to be deleted.

Returns

{:ok, %Cashtrail.Users.User{}} in case of success.

{:error, %Ecto.Changeset{}} in case of error.

Examples

iex> delete_user(user)
{:ok, %User{}}

iex> delete_user(user)
{:error, %Ecto.Changeset{}}

get_user!(id)

get_user!(Ecto.UUID.t() | String.t()) :: user()

Gets a single user.

Raises Ecto.NoResultsError if the User does not exist.

See Cashtrail.Users.User to have more detailed info about the returned struct.

Expected Arguments

id - A string that is the unique id of the user to be found.

Examples

iex> get_user!(123)
%User{}

iex> get_user!(456)
** (Ecto.NoResultsError)

get_user_by(params \\ [])

get_user_by(keyword() | map()) :: nil | user()

Gets a single user by the given param.

Returns nil the User does not exist.

See Cashtrail.Users.User to have more detailed info about the returned struct and the params attributes that can be given.

Expected Arguments

params - A keyword or a map with the attributes of the user to be found.

Examples

iex> get_user_by(email: "john@example.com")
%User{}

iex> get_user_by(email: "noexists')
nil

list_users(options \\ [])

list_users(keyword()) :: Cashtrail.Paginator.Page.t(user())

Returns a %Cashtrail.Paginator.Page{} struct with a list of users in the :entries field.

Expected arguments

options - A keyword list of the following options:

:search - search users by :first_name, :last_name or :email.

See Cashtrail.Paginator.paginate/2 to see the paginations options.

See Cashtrail.Users.User to have more detailed info about the fields to be filtered or searched.

Examples

iex> list_users()
%Cashtrail.Paginator{entries: [%User{}, ...]}

iex> list_users(search: "my")
%Cashtrail.Paginator{entries: [%User{first_name: "My name"}, ...]}

update_user(user, attrs)

update_user(user(), map()) :: {:ok, user()} | {:error, Ecto.Changeset.t(user())}

Updates a user.

Expected Arguments

user - The %Cashtrail.Users.User{} to be updated.

params - A map with the field of the user to be updated. See create_user/2 to know about the params that can be given.

Returns

{:ok, %Cashtrail.Users.User{}} in case of success.

{:error, %Ecto.Changeset{}} in case of error.

Examples

iex> update_user(user, %{field: new_value})
{:ok, %User{}}

iex> update_user(user, %{field: bad_value})
{:error, %Ecto.Changeset{}}